+++ src/services/index.ts
@@ -38,3 +38,5 @@ export * from "./admin.service";
 export * from "./bid.service";
 export * from "./contact.service";
 export * from "./media.service";
+export * from "./blog.service";
+export * from "./seller.service";